refactor(game): rename 'game-card' component to 'game-link'
