+++ src/entities/game/ui/game-link.vue
