+++ src/entities/game/ui/index.ts
@@ -1,2 +1,2 @@
 export { default as StartGameCard } from './start-game-card.vue';
-export { default as GameCard } from './game-card.vue';
+export { default as GameLink } from './game-link.vue';
